Linted main js. Divide helper-functions and dom elements in different files.
