+++ src/index.html
@@ -93,7 +93,6 @@
         </div>
       </div>
     </div>
-    <script type="module" src="./elements.js"></script>
     <script type="module" src="./main.js"></script>
   </body>
 </html>
