+++ src/main.js
@@ -1,6 +1,13 @@
+/* eslint-disable import/extensions */
+/* eslint-disable no-alert */
+
+
 import {
   myLibrary, tableBody, formButton, form,
 } from './elements.js';
+import {
+  formReset, showForm, errors, checkIfValid,
+} from './helpers.js';
 
 function Book(title, author, pages, read) {
   this.title = title;
@@ -10,9 +17,6 @@ function Book(title, author, pages, read) {
   this.info = () => `${this.title} by ${this.author} with ${this.pages} pages, ${this.read}`;
 }
 
-function showForm() {
-  form.classList.toggle('hidden');
-}
 
 function render() {
   let count = 0;
@@ -42,41 +46,6 @@ function render() {
 
 render();
 
-function checkIfValid(object) {
-  let checked = false;
-  myLibrary.forEach((book) => {
-    const regex = new RegExp(book.title, 'gi');
-    if (object.title.match(regex) !== null) {
-      checked = true;
-    }
-  });
-  return checked;
-}
-
-const formReset = () => {
-  const inputs = form.querySelectorAll('input');
-  inputs.forEach((i) => i.classList.remove('is-valid'));
-  form.reset();
-};
-
-const errors = (input) => {
-  if (input.type !== 'checkbox') {
-    const small = input.parentNode.querySelector('small');
-
-    if (input.value === '') {
-      input.classList.add('is-invalid');
-      small.innerText = `The ${input.placeholder} can't be empty`;
-      small.classList.add('text-danger');
-      return false;
-    }
-    input.classList.add('is-valid');
-    input.classList.remove('is-invalid');
-    small.innerText = `The ${input.placeholder} is valid`;
-    small.classList.remove('text-danger');
-    return true;
-  }
-  return '';
-};
 
 function addBook(e) {
   e.preventDefault();
